+++ cpp/src/groupby/sort/group_quantiles.cu
@@ -23,6 +23,9 @@
 
 #include <groupby/sort/group_reductions.hpp>
 #include <quantiles/quantiles_util.hpp>
+#include <rmm/cuda_stream_view.hpp>
+#include <rmm/device_vector.hpp>
+#include <rmm/exec_policy.hpp>
 
 #include <rmm/thrust_rmm_allocator.h>
 #include <thrust/for_each.h>
@@ -100,7 +103,7 @@ struct quantiles_functor {
     // For each group, calculate quantile
     if (!cudf::is_dictionary(values.type())) {
       auto values_iter = values_view->begin<T>();
-      thrust::for_each_n(rmm::exec_policy(stream)->on(stream.value()),
+      thrust::for_each_n(rmm::exec_policy(stream),
                          thrust::make_counting_iterator(0),
                          num_groups,
                          calculate_quantile_fn<ResultType, decltype(values_iter)>{
@@ -113,7 +116,7 @@ struct quantiles_functor {
                            interpolation});
     } else {
       auto values_iter = cudf::dictionary::detail::make_dictionary_iterator<T>(*values_view);
-      thrust::for_each_n(rmm::exec_policy(stream)->on(stream.value()),
+      thrust::for_each_n(rmm::exec_policy(stream),
                          thrust::make_counting_iterator(0),
                          num_groups,
                          calculate_quantile_fn<ResultType, decltype(values_iter)>{